split class from run_test.py
