+++ run_test.py
@@ -5,31 +5,12 @@ import time
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
-from dataclasses import dataclass
 from config import CHROME_DRIVER_PATH
+from model.buildings import Building
 
 # 데이터 클래스 설정
 general_data = []
 specific_data = []
-
-
-@dataclass
-class Building:
-    building_name: str
-    building_type: str
-    # contract_type : str
-    trading_num: int
-    tenant_num: int
-    wolse_num: int
-    # building_dong : str
-    # exclusive_area : float
-    # shared_area : float
-    # target_floor : int
-    # max_floor : int
-    # window_side : str
-    # trading_price_min : int
-    # trading_price_max : int
-
 
 url = "https://m.land.naver.com/"
